ticket:372 - '''REVERSED''' as changes answers in the offline RA3 Nesting suite. Reversed r9239 of /casim/trunk
git-svn-id: https://metomi/svn/monc.xm/casim/trunk@9241 0f676ef4-b20c-4647-9485-21614760d15f

+++ src/special.F90
@@ -200,7 +200,7 @@ contains
       end do
       f=cc+f*z*(2.0-4.0*cc)/sqrt(pi)
     else
-      cc=cc*int(abs(z)/z)
+      cc=cc*abs(z)/z
       j=3+int(32/abs(z))
       f=0.0
       do while(j /= 0)
